Add support for provider networks
Quantum lost the ability to configure provider
networks between folsom and grizzly in part
due to changes made in commit 094c2ae:
only set network_vlan_ranges for vlan network types

Since network_vlan_ranges informs the plugin
which provider networks are available, it is
still needed even when not in vlan mode.

This patch changes the behavior to be:

- if the user sets vlan_ranges, the config option
will be set regardless of plugin

- if the user doesn't set it, a default is used
when in vlan mode, while the value is ensured
absent when in gre mode.

Change-Id: I1638d28da87041798c0b183923137c3bd535c6c3